+++ packages/engine/src/index.ts
@@ -1,5 +1,6 @@
 export type * from "./types";
 export * from "./types/poll";
 export * from "./types/channel";
-export * from "./questionTypes";
+export * from "./questionTypeStore";
+export * from "./themeStore";
 export * from "./ranking-systems";
