+++ packages/form-schema/src/index.ts
@@ -1,3 +1,3 @@
 export type * from "./types/form";
-export type * from "./types/field-types";
+export * from "./types/field-types";
 export { schemaToDefaultValues } from "./schema/schemaToDefaultValues";
