+++ packages/form/src/index.ts
@@ -2,6 +2,4 @@ export * from "./base-form/FormFields";
 export { schemaToZod } from "./schema/schemaToZod";
 export { SchemaForm, getFormId } from "./SchemaForm";
 
-export { pluginField } from "./fields/PluginField";
-
 import "./fields/register";
