+++ packages/ui/src/index.ts
@@ -1,6 +1,5 @@
 export * from "./components/";
 
-export { themeStore } from "./theming/themeStore";
 export { ThemeProvider } from "./theming/ThemeContext";
 
 import "./registerThemes";
