+++ packages/ui/src/registerThemes.ts
@@ -1,4 +1,4 @@
-import { themeStore } from "./theming/themeStore";
+import { themeStore } from "@marcianosrs/engine";
 import { htmlTheme } from "./themes/html/theme";
 import { winterTheme } from "./themes/winter/theme";
