--- packages/ui/src/theming/themeStore.tsx
@@ -1,7 +1,0 @@
-import { createPluginStore } from "@marcianosrs/utils";
-import { ThemePlugin } from "./types";
-import { TypedForm } from "@marcianosrs/form-schema";
-
-export const themeStore = createPluginStore<ThemePlugin<TypedForm>>(
-    (plugin) => plugin.name
-);
