+++ packages/ui/src/theming/useThemedElement.ts
@@ -1,12 +1,16 @@
+import { ThemePlugin, themeStore } from "@marcianosrs/engine";
 import { htmlTheme } from "../themes/html/theme";
 import { useTheme } from "./ThemeContext";
 import type { Theme } from "./ThemeType";
-import { themeStore } from "./themeStore";
+import { TypedForm } from "@marcianosrs/form-schema";
 
 export const useThemedElement = <TElement extends keyof Theme>(
-    elementType: TElement
+	elementType: TElement
 ): Theme[TElement] => {
-    const { themeName } = useTheme();
-    const plugin = themeStore.get(themeName) ?? htmlTheme;
-    return plugin.components[elementType];
+	const { themeName } = useTheme();
+	const plugin = (themeStore.get(themeName) ?? htmlTheme) as ThemePlugin<
+		TypedForm,
+		Theme
+	>;
+	return plugin.components[elementType];
 };
